+++ jiant/tasks/lib/edge_probing/nonterminal.py
@@ -1,65 +1,81 @@
+"""Constituent labeling (assigning a non-terminal label) Edge Probing task.
+
+Task source paper: https://arxiv.org/pdf/1905.06316.pdf.
+Task data prep directions: https://github.com/nyu-mll/jiant/blob/master/probing/data/README.md.
+
+"""
 from dataclasses import dataclass
 
 from jiant.tasks.lib.templates.shared import labels_to_bimap
-from jiant.tasks.lib.templates import edge_probing_two_span
+from jiant.tasks.lib.templates import edge_probing_single_span
 from jiant.utils.python.io import read_json_lines
 
 
 @dataclass
-class Example(edge_probing_two_span.Example):
+class Example(edge_probing_single_span.Example):
     @property
     def task(self):
-        return Spr2Task
+        return NonterminalTask
 
 
 @dataclass
-class TokenizedExample(edge_probing_two_span.TokenizedExample):
+class TokenizedExample(edge_probing_single_span.TokenizedExample):
     pass
 
 
 @dataclass
-class DataRow(edge_probing_two_span.DataRow):
+class DataRow(edge_probing_single_span.DataRow):
     pass
 
 
 @dataclass
-class Batch(edge_probing_two_span.Batch):
+class Batch(edge_probing_single_span.Batch):
     pass
 
 
-class Spr2Task(edge_probing_two_span.AbstractProbingTask):
+class NonterminalTask(edge_probing_single_span.AbstractProbingTask):
     Example = Example
     TokenizedExample = TokenizedExample
     DataRow = DataRow
     Batch = Batch
 
     LABELS = [
-        "awareness",
-        "change_of_location",
-        "change_of_possession",
-        "change_of_state",
-        "change_of_state_continuous",
-        "changes_possession",
-        "existed_after",
-        "existed_before",
-        "existed_during",
-        "exists_as_physical",
-        "instigation",
-        "location_of_event",
-        "makes_physical_contact",
-        "partitive",
-        "predicate_changed_argument",
-        "sentient",
-        "stationary",
-        "volition",
-        "was_for_benefit",
-        "was_used",
+        "ADJP",
+        "ADVP",
+        "CONJP",
+        "EMBED",
+        "FRAG",
+        "INTJ",
+        "LST",
+        "META",
+        "NAC",
+        "NML",
+        "NP",
+        "NX",
+        "PP",
+        "PRN",
+        "PRT",
+        "QP",
+        "RRC",
+        "S",
+        "SBAR",
+        "SBARQ",
+        "SINV",
+        "SQ",
+        "TOP",
+        "UCP",
+        "VP",
+        "WHADJP",
+        "WHADVP",
+        "WHNP",
+        "WHPP",
+        "X",
     ]
     LABEL_TO_ID, ID_TO_LABEL = labels_to_bimap(LABELS)
 
     @property
     def num_spans(self):
-        return 2
+        return 1
 
     def get_train_examples(self):
         return self._create_examples(lines=read_json_lines(self.train_path), set_type="train")
@@ -75,15 +91,13 @@ class Spr2Task(edge_probing_two_span.AbstractProbingTask):
         examples = []
         for (line_num, line) in enumerate(lines):
             for (target_num, target) in enumerate(line["targets"]):
-                span1 = target["span1"]
-                span2 = target["span2"]
+                span = target["span1"]
                 examples.append(
                     Example(
                         guid="%s-%s-%s" % (set_type, line_num, target_num),
                         text=line["text"],
-                        span1=span1,
-                        span2=span2,
-                        labels=target["label"] if set_type != "test" else [cls.LABELS[-1]],
+                        span=span,
+                        labels=[target["label"]] if set_type != "test" else [cls.LABELS[-1]],
                     )
                 )
         return examples
