+++ jiant/tasks/retrieval.py
@@ -10,10 +10,12 @@ from jiant.tasks.lib.ccg import CCGTask
 from jiant.tasks.lib.cola import ColaTask
 from jiant.tasks.lib.commitmentbank import CommitmentBankTask
 from jiant.tasks.lib.commonsenseqa import CommonsenseQATask
+from jiant.tasks.lib.edge_probing.nonterminal import NonterminalTask
 from jiant.tasks.lib.copa import CopaTask
-from jiant.tasks.lib.coref import CorefTask
+from jiant.tasks.lib.edge_probing.coref import CorefTask
 from jiant.tasks.lib.cosmosqa import CosmosQATask
-from jiant.tasks.lib.dpr import DprTask
+from jiant.tasks.lib.edge_probing.dep import DepTask
+from jiant.tasks.lib.edge_probing.dpr import DprTask
 from jiant.tasks.lib.glue_diagnostics import GlueDiagnosticsTask
 from jiant.tasks.lib.hellaswag import HellaSwagTask
 from jiant.tasks.lib.mlm_crosslingual_wiki import MLMCrosslingualWikiTask
@@ -22,22 +24,23 @@ from jiant.tasks.lib.mlqa import MlqaTask
 from jiant.tasks.lib.mnli import MnliTask
 from jiant.tasks.lib.mrpc import MrpcTask
 from jiant.tasks.lib.multirc import MultiRCTask
-from jiant.tasks.lib.ner import NerTask
+from jiant.tasks.lib.edge_probing.ner import NerTask
 from jiant.tasks.lib.panx import PanxPreprocTask
 from jiant.tasks.lib.pawsx import PawsXTask
+from jiant.tasks.lib.edge_probing.pos import PosTask
 from jiant.tasks.lib.qqp import QqpTask
 from jiant.tasks.lib.qnli import QnliTask
 from jiant.tasks.lib.record import ReCoRDTask
 from jiant.tasks.lib.rte import RteTask
 from jiant.tasks.lib.scitail import SciTailTask
 from jiant.tasks.lib.senteval.tense import SentevalTenseTask
-from jiant.tasks.lib.semeval import SemevalTask
+from jiant.tasks.lib.edge_probing.semeval import SemevalTask
 from jiant.tasks.lib.snli import SnliTask
 from jiant.tasks.lib.socialiqa import SocialIQATask
-from jiant.tasks.lib.spr1 import Spr1Task
-from jiant.tasks.lib.spr2 import Spr2Task
+from jiant.tasks.lib.edge_probing.spr1 import Spr1Task
+from jiant.tasks.lib.edge_probing.spr2 import Spr2Task
 from jiant.tasks.lib.squad import SquadTask
-from jiant.tasks.lib.srl import SrlTask
+from jiant.tasks.lib.edge_probing.srl import SrlTask
 from jiant.tasks.lib.sst import SstTask
 from jiant.tasks.lib.stsb import StsbTask
 from jiant.tasks.lib.superglue_axg import SuperglueWinogenderDiagnosticsTask
@@ -67,9 +70,11 @@ TASK_DICT = {
     "ccg": CCGTask,
     "cola": ColaTask,
     "commonsenseqa": CommonsenseQATask,
+    "nonterminal": NonterminalTask,
     "copa": CopaTask,
     "coref": CorefTask,
     "cosmosqa": CosmosQATask,
+    "dep": DepTask,
     "dpr": DprTask,
     "glue_diagnostics": GlueDiagnosticsTask,
     "hellaswag": HellaSwagTask,
@@ -82,6 +87,7 @@ TASK_DICT = {
     "ner": NerTask,
     "pawsx": PawsXTask,
     "panx": PanxPreprocTask,
+    "pos": PosTask,
     "qnli": QnliTask,
     "qqp": QqpTask,
     "record": ReCoRDTask,
